добавил абстрактный класс component
